"factored out PWM and LEDIndicators from HAL; updates #1173"
