+++ software/controller/lib/core/actuators.cpp
@@ -27,7 +27,7 @@ static PinchValve exhale_pinch(1, "exhale", " for exhale valve");
 
 void ActuatorsExecute(const ActuatorsState &desired_state) {
   // set blower PWM
-  hal.AnalogWrite(PwmPin::Blower, desired_state.blower_power);
+  hal.pwm_.set(PwmPin::Blower, desired_state.blower_power);
 
   // Set the blower pinch valve position
   if (desired_state.blower_valve)
